qlskapp/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.1
cryptography==44.0.3
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.7.0
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
dotenv==0.9.9
drf-yasg==1.21.10
fonttools==4.56.0
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
kiwisolver==1.4.8
matplotlib==3.10.1
mysqlclient==2.2.7
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
seaborn==0.13.2
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
